+++ Pods/Target Support Files/Alamofire-watchOS2.0/Alamofire-watchOS2.0-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_Alamofire_watchOS2_0 : NSObject
-@end
-@implementation PodsDummy_Alamofire_watchOS2_0
-@end