+++ Pods/Target Support Files/Kommander-iOS/Kommander-iOS-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_Kommander_iOS : NSObject
+@end
+@implementation PodsDummy_Kommander_iOS
+@end