+++ Pods/Target Support Files/Kommander-iOS8.0/Kommander-iOS8.0-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_Kommander_iOS8_0 : NSObject
-@end
-@implementation PodsDummy_Kommander_iOS8_0
-@end