+++ Pods/Target Support Files/Kommander-macOS/Kommander-macOS-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_Kommander_macOS : NSObject
+@end
+@implementation PodsDummy_Kommander_macOS
+@end