+++ Pods/Target Support Files/Kommander-macOS10.10/Kommander-macOS10.10-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_Kommander_macOS10_10 : NSObject
-@end
-@implementation PodsDummy_Kommander_macOS10_10
-@end