+++ Pods/Target Support Files/Kommander-watchOS/Kommander-watchOS-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_Kommander_watchOS : NSObject
+@end
+@implementation PodsDummy_Kommander_watchOS
+@end