+++ Pods/Target Support Files/Moya-watchOS/Moya-watchOS-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_Moya_watchOS : NSObject
+@end
+@implementation PodsDummy_Moya_watchOS
+@end