+++ Pods/Target Support Files/Result-watchOS2.0/Result-watchOS2.0-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_Result_watchOS2_0 : NSObject
-@end
-@implementation PodsDummy_Result_watchOS2_0
-@end