+++ Pods/Target Support Files/RxSwift-watchOS2.0/RxSwift-watchOS2.0-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_RxSwift_watchOS2_0 : NSObject
-@end
-@implementation PodsDummy_RxSwift_watchOS2_0
-@end